@@ -7,7 +7,7 @@ homepage: https://flyer.chat
 repository: https://github.com/flyerhq/flutter_firebase_chat_core
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
